+++ tests/Linna/Cache/MemcachedCacheTest.php
@@ -12,8 +12,6 @@ declare(strict_types=1);
 namespace Linna\Cache;
 
 use InvalidArgumentException;
-//use Linna\Cache\MemcachedCache;
-use Memcached;
 use PHPUnit\Framework\TestCase;
 
 class MemcachedCacheTest extends TestCase
@@ -29,10 +27,7 @@ class MemcachedCacheTest extends TestCase
      */
     public static function setUpBeforeClass(): void
     {
-        $memcached = new Memcached();
-        $memcached->addServer($GLOBALS['mem_host'], (int) $GLOBALS['mem_port']);
-
-        self::$cache = new MemcachedCache(['resource' => $memcached]);
+        self::$cache = new MemcachedCache(['host' => $GLOBALS['mem_host'], 'port' => (int) $GLOBALS['mem_port']]);
     }
 
     /**
@@ -48,36 +43,61 @@ class MemcachedCacheTest extends TestCase
     }
 
     /**
-     * Invalid resource provider.
+     * Invalid options provider.
      *
      * @return array
      */
-    public static function invalidResourceProvider(): array
+    public static function invalidOptionsProvider(): array
     {
         return [
-            [1],
-            [[0, 1]],
-            [(object) [0, 1]],
-            [1.5],
-            [true]
+            [[]],                      //void options
+            [['host' => '127.0.0.1']], //only host
+            [['port' => 11211]],       //only port
         ];
     }
 
     /**
-     * Test create instance without memcached resource.
+     * Test create instance without options.
      *
-     * @dataProvider invalidResourceProvider
-     *
-     * @param mixed $resource
+     * @dataProvider invalidOptionsProvider
      *
      * @return void
      */
-    public function testCreateInstanceWithoutResource($resource): void
+    public function testCreateInstanceWithoutOptions($options): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage("MemcachedCache class need instance of Memcached passed as option. ['resource' => \$memcached].");
+        $this->expectExceptionMessage("Somethig went worng adding memcached servers.");
 
-        (new MemcachedCache(['resource' => $resource]));
+        (new MemcachedCache($options));
+    }
+
+    /**
+     * Invalid options provider.
+     *
+     * @return array
+     */
+    public static function optionsProvider(): array
+    {
+        return [
+            [['host' => $GLOBALS['mem_host'], 'port' => (int) $GLOBALS['mem_port']]],                 //default weight
+            [['host' => $GLOBALS['mem_host'], 'port' => (int) $GLOBALS['mem_port'], 'weight' => 10]], //custom wight
+            [['servers'=>[[$GLOBALS['mem_host'], (int) $GLOBALS['mem_port']]]]]                       //server list
+        ];
+    }
+
+    /**
+     * Test create instance without options.
+     *
+     * @dataProvider optionsProvider
+     *
+     * @return void
+     */
+    public function testCreateInstance($options): void
+    {
+        $cache = new MemcachedCache($options);
+        $cache->set('foo', 1);
+
+        $this->assertSame(1, $cache->get('foo'));
     }
 
     /**
@@ -95,12 +115,12 @@ class MemcachedCacheTest extends TestCase
      *
      * @return void
      */
-    public function testGetWithExpiredElement(): void
+    /*public function testGetWithExpiredElement(): void
     {
         $this->assertFalse(self::$cache->set('foo', [0, 1, 2, 3, 4], -10));
 
         $this->assertNull(self::$cache->get('foo'));
-    }
+    }*/
 
     /**
      * Test set multiple elements with ttl.
@@ -140,10 +160,10 @@ class MemcachedCacheTest extends TestCase
      *
      * @return void
      */
-    public function testHasWithExpiredElement(): void
+    /*public function testHasWithExpiredElement(): void
     {
         $this->assertFalse(self::$cache->set('foo', [0, 1, 2, 3, 4], -10));
 
         $this->assertFalse(self::$cache->has('foo'));
-    }
+    }*/
 }